refactor postWithBoardsEmbed
